chore: rename internally EditableHeading to LegacyEditableHeading
