+++ src/components/Modal/__stories__/modal.stories.mdx
@@ -12,7 +12,7 @@ import {
 } from "../../../storybook/components/related-components/component-description-map";
 import Modal from "../Modal";
 import { DialogAsModalBadExample, ModalExampleWrapper, useHelperOpenModalButton } from "./helpers";
-import { EditableHeading } from "../../../components";
+import { LegacyEditableHeading } from "../../../components";
 import ModalFooterButtons from "../ModalFooter/ModalFooterButtons/ModalFooterButtons";
 import { Upgrade } from "../../Icon/Icons";
 
@@ -410,7 +410,7 @@ clicking outside. Use this variant in case of sensitive or important messages, a
             {/** Please follow the design guidelines when implementing custom title **/}
             <ModalHeader
               description={"Description text goes here"}
-              title={<EditableHeading type={EditableHeading.types.h2} placeholder={"Modal title"} />}
+              title={<LegacyEditableHeading type={LegacyEditableHeading.types.h2} placeholder={"Modal title"} />}
             />
             <ModalContent>Modal content goes here</ModalContent>
             <ModalFooterButtons
