+++ src/components/index.js
@@ -44,7 +44,7 @@ export { default as HiddenText } from "./HiddenText/HiddenText";
 export { default as ExpandCollapse } from "./ExpandCollapse/ExpandCollapse";
 export { default as Heading } from "./LegacyHeading/LegacyHeading";
 export { default as EditableInput } from "./EditableInput/EditableInput";
-export { default as EditableHeading } from "./EditableHeading/EditableHeading";
+export { default as EditableHeading } from "./LegacyEditableHeading/LegacyEditableHeading";
 export { default as ResponsiveList } from "./ResponsiveList/ResponsiveList";
 export { default as BreadcrumbsBar } from "./BreadcrumbsBar/BreadcrumbsBar";
 export { default as BreadcrumbItem } from "./BreadcrumbsBar/BreadcrumbItem/BreadcrumbItem";
